Update timer.js
The main modifications are:
We introduce an animationFrameId variable to keep track of the animation frame.
In the stopCountdown function, we cancel the animation frame to stop updates.
We modify the visibilitychange event listener to use setTimeout when the page is hidden and switch back to requestAnimationFrame when it becomes visible again.
These changes should keep the Picture-in-Picture window active and updating even when you switch to another tab. The PiP window should now float on top of other tabs, similar to other video player PiP implementations.